Refactor Time Entry validation
