+++ gourmet/po/POTFILES.in
@@ -54,9 +54,8 @@ gtk_extras/numberEntry.py
 gtk_extras/optionTable.py
 gtk_extras/pageable_store.py
 gtk_extras/ratingWidget.py
-gtk_extras/time_entry.py
 gtk_extras/treeview_extras.py
-gtk_extras/validatingEntry.py
+gtk_extras/validation.py
 importers/__init__.py
 importers/generic_recipe_parser.py
 importers/html_importer.py
